Breaks packages down
