+++ configuration/cluster/hyperkube/hyperkube.go
@@ -1,0 +1,9 @@
+// Package hyperkube provides configuration structures for Hyperkube.
+package hyperkube
+
+// Hyperkube holds configuration for Hyperkube settings.
+type Hyperkube struct {
+	// Version is the version of Hyperkube.
+	// e.g: 'v1.5.2_coreos.0'
+	Version string
+}
